remove region syncer and region storage
