+++ proto/proto/pdpb.proto
@@ -55,8 +55,6 @@ service PD {
 
     rpc UpdateGCSafePoint(UpdateGCSafePointRequest) returns (UpdateGCSafePointResponse) {}
 
-    rpc SyncRegions(stream SyncRegionRequest) returns (stream SyncRegionResponse) {}
-
     rpc GetOperator(GetOperatorRequest) returns (GetOperatorResponse) {}
 }
 
@@ -475,22 +473,6 @@ message UpdateGCSafePointResponse {
     ResponseHeader header = 1;
 
     uint64 new_safe_point = 2;
-}
-
-message SyncRegionRequest{
-    RequestHeader header = 1;
-    Member member = 2;
-    // the follower PD will use the start index to locate historical changes
-    // that require synchronization.
-    uint64 start_index = 3;
-}
-
-message SyncRegionResponse{
-    ResponseHeader header = 1;
-    // the leader PD will send the repsonds include
-    // changed regions records and the index of the first record.
-    repeated metapb.Region regions = 2;
-    uint64 start_index = 3;
 }
 
 message GetOperatorRequest {
