+++ scheduler/server/cluster.go
@@ -29,7 +29,6 @@ import (
 	"github.com/pingcap-incubator/tinykv/scheduler/server/config"
 	"github.com/pingcap-incubator/tinykv/scheduler/server/core"
 	"github.com/pingcap-incubator/tinykv/scheduler/server/id"
-	syncer "github.com/pingcap-incubator/tinykv/scheduler/server/region_syncer"
 	"github.com/pingcap-incubator/tinykv/scheduler/server/schedule"
 	"github.com/pingcap/errcode"
 	"github.com/pingcap/log"
@@ -74,7 +73,6 @@ type RaftCluster struct {
 
 	wg           sync.WaitGroup
 	quit         chan struct{}
-	regionSyncer *syncer.RegionSyncer
 }
 
 // ClusterStatus saves some state information
@@ -90,7 +88,6 @@ func newRaftCluster(ctx context.Context, s *Server, clusterID uint64) *RaftClust
 		running:      false,
 		clusterID:    clusterID,
 		clusterRoot:  s.getClusterRootPath(),
-		regionSyncer: syncer.NewRegionSyncer(s),
 	}
 }
 
@@ -164,10 +161,9 @@ func (c *RaftCluster) start() error {
 	c.coordinator = newCoordinator(c.ctx, cluster, c.s.hbStreams)
 	c.quit = make(chan struct{})
 
-	c.wg.Add(3)
+	c.wg.Add(2)
 	go c.runCoordinator()
 	go c.runBackgroundJobs(backgroundJobInterval)
-	go c.syncRegions()
 	c.running = true
 
 	return nil
@@ -234,12 +230,6 @@ func (c *RaftCluster) runCoordinator() {
 	c.coordinator.run()
 	<-c.coordinator.ctx.Done()
 	log.Info("coordinator is stopping")
-}
-
-func (c *RaftCluster) syncRegions() {
-	defer logutil.LogPanic()
-	defer c.wg.Done()
-	c.regionSyncer.RunServer(c.changedRegionNotifier(), c.quit)
 }
 
 func (c *RaftCluster) stop() {
