--- scheduler/server/region_syncer/metrics.go
@@ -1,28 +1,0 @@
-// Copyright 2018 PingCAP, Inc.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package syncer
-
-import "github.com/prometheus/client_golang/prometheus"
-
-var regionSyncerStatus = prometheus.NewGaugeVec(
-	prometheus.GaugeOpts{
-		Namespace: "pd",
-		Subsystem: "region_syncer",
-		Name:      "status",
-		Help:      "Inner status of the region syncer.",
-	}, []string{"type"})
-
-func init() {
-	prometheus.MustRegister(regionSyncerStatus)
-}
